Made the separation for tinyfs.
